file_size type changed

+++ trace-selection/datatype.h
@@ -42,7 +42,7 @@ typedef struct {
     unsigned int nb_blocks;
     unsigned int size;
     unsigned int trace_count;
-    unsigned int file_size;
+    unsigned long long file_size;
     // FILE* file; // We will no longer need this
     unsigned char *mapped_file; // base pointer to the mmapped file
     GHashTable *hash_table; // Hash table for storing the traces structure
